+++ relay-server/src/processing/utils/attachments.rs
@@ -40,6 +40,8 @@ fn validate(item: &Item, config: &Config) -> Result<(), ProcessingError> {
 
     #[cfg(feature = "processing")]
     {
+        use crate::services::upload::{Final, SignedLocation};
+
         if !item.is_attachment_ref() {
             return Ok(());
         }
@@ -47,16 +49,13 @@ fn validate(item: &Item, config: &Config) -> Result<(), ProcessingError> {
         let payload = item.payload();
         let payload: AttachmentPlaceholder =
             serde_json::from_slice(&payload).map_err(|_| ProcessingError::InvalidAttachmentRef)?;
-        let signed_location =
-            crate::services::upload::SignedLocation::try_from_str(payload.location)
-                .ok_or(ProcessingError::InvalidAttachmentRef)?;
+        let signed_location: SignedLocation<Final> = SignedLocation::try_from_str(payload.location)
+            .ok_or(ProcessingError::InvalidAttachmentRef)?;
         // NOTE: Using the received timestamp here breaks tests without a pop-relay.
         let location = signed_location
             .verify(chrono::Utc::now(), config)
             .map_err(|_| ProcessingError::InvalidAttachmentRef)?;
-        let signed_length = location
-            .length
-            .ok_or(ProcessingError::InvalidAttachmentRef)?;
+        let signed_length = location.length.into_inner();
 
         match item.attachment_body_size() == signed_length {
             true => Ok(()),